Move `Api` module to parent directory
Close #49
