+++ src/api/details.rs
@@ -1,6 +1,3 @@
-pub use self::language::Language;
-pub mod language;
-
 pub const SERVER_URL: &str = "baca.ii.uj.edu.pl";
 pub const PERMUTATION: &str = "5A4AE95C27260DF45F17F9BF027335F6";
 pub const EMPTY_RESPONSE: &str = "//OK[0,[],0,7]";
