+++ src/api/mod.rs
@@ -1,0 +1,8 @@
+pub use self::request::Request;
+pub use self::request_type::RequestType;
+
+pub mod baca_api;
+pub mod baca_service;
+pub mod details;
+mod request;
+mod request_type;
