+++ src/api/request.rs
@@ -1,4 +1,4 @@
-use crate::baca::api::RequestType;
+use crate::api::RequestType;
 use crate::error;
 use crate::error::Error;
 use crate::model::Task;
