--- src/baca/api/mod.rs
@@ -1,7 +1,0 @@
-pub mod baca_api;
-pub mod baca_service;
-mod request;
-mod request_type;
-
-pub use self::request::Request;
-pub use self::request_type::RequestType;
