--- src/baca/mod.rs
@@ -1,2 +1,0 @@
-pub mod api;
-pub mod details;
