+++ src/command/details.rs
@@ -1,4 +1,4 @@
-use crate::baca::api::baca_api::BacaApi;
+use crate::api::baca_api::BacaApi;
 use crate::command::Command;
 use crate::error::Result;
 use crate::workspace::Workspace;
@@ -46,7 +46,7 @@ impl Command for Details {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::baca::api::baca_api::MockBacaApi;
+    use crate::api::baca_api::MockBacaApi;
     use crate::workspace::{InstanceData, MockWorkspace};
 
     #[test]
