+++ src/command/last.rs
@@ -1,4 +1,4 @@
-use crate::baca::api::baca_api::BacaApi;
+use crate::api::baca_api::BacaApi;
 use crate::command::details::Details;
 use crate::command::Command;
 use crate::error;
@@ -30,7 +30,7 @@ impl Command for Last {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::baca::api::baca_api::MockBacaApi;
+    use crate::api::baca_api::MockBacaApi;
     use crate::model::SubmitStatus;
     use crate::model::{Results, Submit};
     use crate::workspace::{InstanceData, MockWorkspace};
