+++ src/command/mod.rs
@@ -1,4 +1,4 @@
-use crate::baca::api::baca_api::BacaApi;
+use crate::api::baca_api::BacaApi;
 use crate::command::details::Details;
 use crate::command::init::Init;
 use crate::command::last::Last;
