+++ src/model/mod.rs
@@ -1,17 +1,14 @@
+mod language;
 mod results;
-pub use self::results::Results;
-
-mod submit_status;
-pub use self::submit_status::SubmitStatus;
-
 mod submit;
-pub use self::submit::Submit;
-
-mod tasks;
-pub use self::tasks::Tasks;
-
+mod submit_status;
 mod task;
-pub use self::task::Task;
-
+mod tasks;
 mod test_results;
+pub use self::language::Language;
+pub use self::results::Results;
+pub use self::submit::Submit;
+pub use self::submit_status::SubmitStatus;
+pub use self::task::Task;
+pub use self::tasks::Tasks;
 pub use self::test_results::TestResults;
